Move getStringEventTagFromId back into SecurityLogsFragment
